MovieListFragment manages menu
